+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add new flavour with workflows and makefile for cluster app
+
 ## [5.19.0] - 2023-02-21
 
 ### Changed