+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- repo setup: better select checks required for PR merge
+
 ## [5.19.0] - 2023-02-21
 
 ### Changed