+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
 - Add new flavour with workflows and makefile for cluster apps.
+
+### Changed
+
+- repo setup: better select checks required for PR merge
 
 ## [5.19.0] - 2023-02-21