+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `--dry-run` flag to `devctl repo setup` command.
+
+### Fixed
+
 - Add new flavour with workflows and makefile for cluster apps.
 
 ### Changed