+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [5.20.0] - 2023-03-22
 
+## [5.20.0] - 2023-03-22
+
 ### Added
 
 - Add `--dry-run` flag to `devctl repo setup` command.
@@ -609,6 +611,7 @@ Renovate config
  - First release.
 
 [Unreleased]: https://github.com/giantswarm/devctl/compare/v5.20.0...HEAD
+[5.20.0]: https://github.com/giantswarm/devctl/compare/v5.20.0...v5.20.0
 [5.20.0]: https://github.com/giantswarm/devctl/compare/v5.19.0...v5.20.0
 [5.19.0]: https://github.com/giantswarm/devctl/compare/v5.18.3...v5.19.0
 [5.18.3]: https://github.com/giantswarm/devctl/compare/v5.18.2...v5.18.3